requirements.txt
streamlit
python-dotenv
pypdf
torch
sentence-transformers
langchain
langchain-community
langchain-google-genai
qdrant-client
python-docx
pandas
openpyxl